// packages/examples/user-management/src/generated/resolvers.ts
export type Maybe<T> = T | null | undefined;
// ...
export type Omit<T, K extends keyof T> = Pick<T, Exclude<keyof T, K>>;
export type RequireFields<T, K extends keyof T> = Omit<T, K> & { [P in K]-?: NonNullable<T[P]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: { input: import("@cqrs-framework/core").AggregateId; output: import("@cqrs-framework/core").AggregateId; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
};

export type CreateUserInput = {
  readonly email: Scalars['String']['input'];
  readonly name: Scalars['String']['input'];
};

export type CreateUserPayload = {
  readonly __typename?: 'CreateUserPayload';
  readonly errors: Maybe<ReadonlyArray<Error>>;
  readonly success: Scalars['Boolean']['output'];
  readonly user: Maybe<User>;
};

export type DeleteUserPayload = {
  readonly __typename?: 'DeleteUserPayload';
  readonly errors: Maybe<ReadonlyArray<Error>>;
  readonly success: Scalars['Boolean']['output'];
};

export type Error = {
  readonly __typename?: 'Error';
  readonly field: Maybe<Scalars['String']['output']>;
  readonly message: Scalars['String']['output'];
};

export type Mutation = {
  readonly __typename?: 'Mutation';
  readonly createUser: CreateUserPayload;
  readonly deleteUser: DeleteUserPayload;
  readonly updateUser: UpdateUserPayload;
};


export type MutationCreateUserArgs = {
  input: CreateUserInput;
};


export type MutationDeleteUserArgs = {
  id: Scalars['ID']['input'];
};


export type MutationUpdateUserArgs = {
  id: Scalars['ID']['input'];
  input: UpdateUserInput;
};

export type Query = {
  readonly __typename?: 'Query';
  readonly _empty: Maybe<Scalars['String']['output']>;
  readonly getUser: Maybe<User>;
  readonly listUsers: UserList;
  readonly searchUsers: ReadonlyArray<User>;
};


export type QueryGetUserArgs = {
  id: Scalars['ID']['input'];
};


export type QueryListUsersArgs = {
  limit?: InputMaybe<Scalars['Int']['input']>;
  offset?: InputMaybe<Scalars['Int']['input']>;
};


export type QuerySearchUsersArgs = {
  query: Scalars['String']['input'];
};

export type UpdateUserInput = {
  readonly email?: InputMaybe<Scalars['String']['input']>;
  readonly name?: InputMaybe<Scalars['String']['input']>;
};

export type UpdateUserPayload = {
  readonly __typename?: 'UpdateUserPayload';
  readonly errors: Maybe<ReadonlyArray<Error>>;
  readonly success: Scalars['Boolean']['output'];
  readonly user: Maybe<User>;
};

export type User = {
  readonly __typename?: 'User';
  readonly createdAt: Scalars['String']['output'];
  readonly email: Scalars['String']['output'];
  readonly id: Scalars['ID']['output'];
  readonly name: Scalars['String']['output'];
  readonly updatedAt: Scalars['String']['output'];
};

export type UserList = {
  readonly __typename?: 'UserList';
  readonly hasMore: Scalars['Boolean']['output'];
  readonly total: Scalars['Int']['output'];
  readonly users: ReadonlyArray<User>;
};



export type ResolverTypeWrapper<T> = Promise<T> | T;


export type ResolverWithResolve<TResult, TParent, TContext, TArgs> = {
  resolve: ResolverFn<TResult, TParent, TContext, TArgs>;
};
export type Resolver<TResult, TParent = {}, TContext = {}, TArgs = {}> = ResolverFn<TResult, TParent, TContext, TArgs> | ResolverWithResolve<TResult, TParent, TContext, TArgs>;

export type ResolverFn<TResult, TParent, TContext, TArgs> = (
  parent: TParent,
  args: TArgs,
  context: TContext,
  info: GraphQLResolveInfo
) => Promise<TResult> | TResult;
// ...
export type ResolversTypes = {
  Boolean: ResolverTypeWrapper<Partial<Scalars['Boolean']['output']>>;
  CreateUserInput: ResolverTypeWrapper<import("../domains/user/models/User").CreateUserInput>;
  CreateUserPayload: ResolverTypeWrapper<Partial<Omit<CreateUserPayload, 'errors' | 'user'> & { errors?: Maybe<ReadonlyArray<ResolversTypes['Error']>>, user?: Maybe<ResolversTypes['User']> }>>;
  DeleteUserPayload: ResolverTypeWrapper<Partial<Omit<DeleteUserPayload, 'errors'> & { errors?: Maybe<ReadonlyArray<ResolversTypes['Error']>> }>>;
  Error: ResolverTypeWrapper<{ field: string | null; message: string; code?: string }>;
  ID: ResolverTypeWrapper<Partial<Scalars['ID']['output']>>;
  Int: ResolverTypeWrapper<Partial<Scalars['Int']['output']>>;
  Mutation: ResolverTypeWrapper<import("../server/server").GraphQLRootValue>;
  Query: ResolverTypeWrapper<import("../server/server").GraphQLRootValue>;
  String: ResolverTypeWrapper<Partial<Scalars['String']['output']>>;
  UpdateUserInput: ResolverTypeWrapper<import("../domains/user/models/User").UpdateUserInput>;
  UpdateUserPayload: ResolverTypeWrapper<Partial<Omit<UpdateUserPayload, 'errors' | 'user'> & { errors?: Maybe<ReadonlyArray<ResolversTypes['Error']>>, user?: Maybe<ResolversTypes['User']> }>>;
  User: ResolverTypeWrapper<import("../domains/user/models/User").User>;
  UserList: ResolverTypeWrapper<Partial<Omit<UserList, 'users'> & { users: ReadonlyArray<ResolversTypes['User']> }>>;
};

/** Mapping between all available schema types and the resolvers parents */
export type ResolversParentTypes = {
  Boolean: Partial<Scalars['Boolean']['output']>;
  CreateUserInput: import("../domains/user/models/User").CreateUserInput;
  CreateUserPayload: Partial<Omit<CreateUserPayload, 'errors' | 'user'> & { errors?: Maybe<ReadonlyArray<ResolversParentTypes['Error']>>, user?: Maybe<ResolversParentTypes['User']> }>;
  DeleteUserPayload: Partial<Omit<DeleteUserPayload, 'errors'> & { errors?: Maybe<ReadonlyArray<ResolversParentTypes['Error']>> }>;
  Error: { field: string | null; message: string; code?: string };
  ID: Partial<Scalars['ID']['output']>;
  Int: Partial<Scalars['Int']['output']>;
  Mutation: import("../server/server").GraphQLRootValue;
  Query: import("../server/server").GraphQLRootValue;
  String: Partial<Scalars['String']['output']>;
  UpdateUserInput: import("../domains/user/models/User").UpdateUserInput;
  UpdateUserPayload: Partial<Omit<UpdateUserPayload, 'errors' | 'user'> & { errors?: Maybe<ReadonlyArray<ResolversParentTypes['Error']>>, user?: Maybe<ResolversParentTypes['User']> }>;
  User: import("../domains/user/models/User").User;
  UserList: Partial<Omit<UserList, 'users'> & { users: ReadonlyArray<ResolversParentTypes['User']> }>;
};
// ...
export type MutationResolvers<ContextType = import("../server/server").GraphQLContext, ParentType extends ResolversParentTypes['Mutation'] = ResolversParentTypes['Mutation']> = {
  createUser?: Resolver<ResolversTypes['CreateUserPayload'], ParentType, ContextType, RequireFields<MutationCreateUserArgs, 'input'>>;
  deleteUser?: Resolver<ResolversTypes['DeleteUserPayload'], ParentType, ContextType, RequireFields<MutationDeleteUserArgs, 'id'>>;
  updateUser?: Resolver<ResolversTypes['UpdateUserPayload'], ParentType, ContextType, RequireFields<MutationUpdateUserArgs, 'id' | 'input'>>;
};